requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.1.0
annotated-types==0.6.0
ansi2html==1.8.0
antiorm==1.2.1
anyio==4.3.0
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
attrs==22.2.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.29.7
Babel==2.12.1
beautifulsoup4==4.12.3
blinker==1.6.2
branca==0.7.0
cachetools==5.3.0
camelcase==0.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.1
comtypes==1.3.1
contourpy==1.0.7
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
datasets==2.17.1
db==0.1.1
db-sqlite3==0.0.1
deap==1.3.3
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==4.1.7
dnspython==2.4.2
docopt==0.6.2
easyocr==1.7.1
executing==2.0.1
filelock==3.12.4
Fiona==1.9.4.post1
Flask==2.2.5
Flask-Cors==4.0.0
flatbuffers==23.1.21
folium==0.14.0
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2023.9.2
future==0.18.3
gast==0.4.0
geopandas==0.8.1
gitdb==4.0.10
GitPython==3.1.32
google-ai-generativelanguage==0.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-generativeai==0.4.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
gTTS==2.5.1
h11==0.14.0
h5py==3.8.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
imutils==0.5.4
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.20.0
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.4
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.12.0rc0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langchain==0.1.9
langchain-community==0.0.24
langchain-core==0.1.26
langsmith==0.1.8
lazy_loader==0.3
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
networkx==3.2
ninja==1.11.1.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==0.28.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==4.1.0
playsound==1.3.0
plotly==5.17.0
pluggy==1.0.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.21.12
psutil==5.9.7
pure-eval==0.2.2
pyarrow==13.0.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.0
PyGetWindow==0.0.9
Pygments==2.16.1
pymongo==4.5.0
Pympler==1.0.1
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.6
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pyproj==3.6.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pyswarms==1.1.0
pytest==7.2.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==306
PyYAML==6.0
pyzmq==25.1.2
referencing==0.30.2
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.2.1
scipy==1.10.1
selenium==4.16.0
sentencepiece==0.2.0
shapely==2.0.1
six==1.16.0
sklearn==0.0.post1
sklearn-extensions==0.0.2
sklearn-genetic==0.2
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.27
sqlparse==0.4.3
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.29.0
streamlit-folium==0.16.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
SwarmPackagePy==1.0.0a5
sympy==1.12
tenacity==8.2.3
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc0
tensorflow-estimator==2.12.0rc0
tensorflow-intel==2.12.0rc0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.9.26
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.14.1
transformers==4.38.1
trio==0.24.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2022.7
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.14
validators==0.21.2
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==2.2.3
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.0.3
xlrd==2.0.1
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.16.2
